+++ src/main/java/edu/depaul/reservations/controller/OrganizationController.java
@@ -111,8 +111,13 @@ public class OrganizationController {
     @PostMapping("/delete/{id}")
     public String delete(@PathVariable(name = "id") final Long id,
             final RedirectAttributes redirectAttributes) {
-        organizationService.delete(id);
-        redirectAttributes.addFlashAttribute(WebUtils.MSG_INFO, WebUtils.getMessage("organization.delete.success"));
+        final String referencedWarning = organizationService.getReferencedWarning(id);
+        if (referencedWarning != null) {
+            redirectAttributes.addFlashAttribute(WebUtils.MSG_ERROR, referencedWarning);
+        } else {
+            organizationService.delete(id);
+            redirectAttributes.addFlashAttribute(WebUtils.MSG_INFO, WebUtils.getMessage("organization.delete.success"));
+        }
         return "redirect:/organizations";
     }
 }
