+++ src/main/java/edu/depaul/reservations/controller/UserTypeController.java
@@ -73,8 +73,13 @@ public class UserTypeController {
     @PostMapping("/delete/{id}")
     public String delete(@PathVariable(name = "id") final Long id,
             final RedirectAttributes redirectAttributes) {
-        userTypeService.delete(id);
-        redirectAttributes.addFlashAttribute(WebUtils.MSG_INFO, WebUtils.getMessage("userType.delete.success"));
+        final String referencedWarning = userTypeService.getReferencedWarning(id);
+        if (referencedWarning != null) {
+            redirectAttributes.addFlashAttribute(WebUtils.MSG_ERROR, referencedWarning);
+        } else {
+            userTypeService.delete(id);
+            redirectAttributes.addFlashAttribute(WebUtils.MSG_INFO, WebUtils.getMessage("userType.delete.success"));
+        }
         return "redirect:/usertypes";
     }
 }
